added API change that comment information no longer accompanies PostDetail
