+++ core/model/src/commonMain/kotlin/me/matsumo/fanbox/core/model/fanbox/entity/FanboxPostCommentItemsEntity.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class FanboxPostCommentItemsEntity(
     @SerialName("body")
-    val body: FanboxPostDetailEntity.Body.CommentList,
+    val body: FanboxCommentsEntity,
 )
